refactor(stage): move index history tests to separate test file
Extract shared test helpers (block creation, provider setup) into a
`common` module and move state indexing tests from `block.rs` to
`index_history.rs`.
