+++ crates/sync/stage/tests/block.rs
@@ -1,30 +1,26 @@
-use std::collections::{BTreeMap, HashMap};
+mod common;
+
+use std::collections::HashMap;
 use std::future::Future;
-use std::ops::RangeInclusive;
 use std::sync::{Arc, Mutex};
 
+use common::{
+    create_provider_with_block_range, create_provider_with_blocks, create_stored_block,
+    get_stored_block_numbers,
+};
 use katana_gateway_client::Client as SequencerGateway;
 use katana_gateway_types::{
     Block, BlockStatus, ConfirmedStateUpdate, StateDiff, StateUpdate, StateUpdateWithBlock,
 };
-use katana_primitives::block::{
-    BlockHash, BlockNumber, FinalityStatus, Header, SealedBlock, SealedBlockWithStatus,
-};
+use katana_primitives::block::{BlockHash, BlockNumber, FinalityStatus};
 use katana_primitives::chain::ChainId;
-use katana_primitives::class::ClassHash;
 use katana_primitives::da::L1DataAvailabilityMode;
-use katana_primitives::state::{StateUpdates, StateUpdatesWithClasses};
-use katana_primitives::version::StarknetVersion;
 use katana_primitives::{felt, ContractAddress, Felt};
-use katana_provider::api::block::{BlockHashProvider, BlockNumberProvider, BlockWriter};
-use katana_provider::api::state::{
-    HistoricalStateRetentionProvider, StateFactoryProvider, StateProvider,
-};
-use katana_provider::api::state_update::StateUpdateProvider;
-use katana_provider::{DbProviderFactory, MutableProvider, ProviderError, ProviderFactory};
+use katana_provider::api::block::BlockNumberProvider;
+use katana_provider::ProviderFactory;
 use katana_stage::blocks::hash::compute_hash;
 use katana_stage::blocks::{BatchBlockDownloader, BlockData, BlockDownloader, Blocks};
-use katana_stage::{PruneInput, Stage, StageExecutionInput};
+use katana_stage::{Stage, StageExecutionInput};
 use katana_tasks::TaskManager;
 use rstest::rstest;
 use starknet::core::types::ResourcePrice;
@@ -130,117 +126,6 @@ impl BlockDownloader for MockBlockDownloader {
 
             Ok(results)
         }
-    }
-}
-
-fn create_provider_with_blocks(blocks: Vec<SealedBlockWithStatus>) -> DbProviderFactory {
-    let provider_factory = DbProviderFactory::new_in_memory();
-    let provider_mut = provider_factory.provider_mut();
-
-    for block in blocks {
-        provider_mut
-            .insert_block_with_states_and_receipts(
-                block,
-                Default::default(),
-                Vec::new(),
-                Vec::new(),
-            )
-            .unwrap();
-    }
-
-    provider_mut.commit().expect("failed to commit");
-    provider_factory
-}
-
-fn create_provider_with_block_range(
-    block_range: RangeInclusive<BlockNumber>,
-    state_updates: BTreeMap<BlockNumber, StateUpdatesWithClasses>,
-) -> DbProviderFactory {
-    let provider_factory = DbProviderFactory::new_in_memory();
-    let provider_mut = provider_factory.provider_mut();
-    let blocks = create_stored_blocks(block_range);
-
-    for block in blocks {
-        let block_num = block.block.header.number;
-        let state_updates = state_updates.get(&block_num).cloned().unwrap_or_default();
-
-        provider_mut
-            .insert_block_with_states_and_receipts(block, state_updates, Vec::new(), Vec::new())
-            .unwrap();
-    }
-
-    provider_mut.commit().expect("failed to commit");
-    provider_factory
-}
-
-fn state_updates_with_contract_changes(
-    contract_address: ContractAddress,
-    class_hash: ClassHash,
-    nonce: Felt,
-    storage_key: Felt,
-    storage_value: Felt,
-) -> StateUpdatesWithClasses {
-    let mut state_updates = StateUpdates::default();
-    state_updates.deployed_contracts.insert(contract_address, class_hash);
-    state_updates.nonce_updates.insert(contract_address, nonce);
-    state_updates
-        .storage_updates
-        .insert(contract_address, BTreeMap::from([(storage_key, storage_value)]));
-
-    StateUpdatesWithClasses { state_updates, ..Default::default() }
-}
-
-/// Gets all stored block numbers from the provider by checking which blocks actually exist.
-fn get_stored_block_numbers(
-    provider: &DbProviderFactory,
-    expected_range: std::ops::RangeInclusive<BlockNumber>,
-) -> Vec<BlockNumber> {
-    let p = provider.provider();
-    expected_range.filter(|&num| p.block_hash_by_num(num).ok().flatten().is_some()).collect()
-}
-
-fn create_stored_blocks(block_range: RangeInclusive<BlockNumber>) -> Vec<SealedBlockWithStatus> {
-    let mut blocks: Vec<SealedBlockWithStatus> = Vec::new();
-
-    let offset = *block_range.start();
-    for i in block_range {
-        let idx = i.abs_diff(offset) as usize;
-        let parent_hash = if idx == 0 { Felt::ZERO } else { blocks[idx - 1].block.hash };
-
-        blocks.push(create_stored_block(i, parent_hash));
-    }
-
-    blocks
-}
-
-fn create_stored_block(block_number: BlockNumber, parent_hash: BlockHash) -> SealedBlockWithStatus {
-    let header = Header {
-        number: block_number,
-        parent_hash,
-        timestamp: 0,
-        sequencer_address: ContractAddress::default(),
-        l1_gas_prices: Default::default(),
-        l1_data_gas_prices: Default::default(),
-        l2_gas_prices: Default::default(),
-        l1_da_mode: L1DataAvailabilityMode::Calldata,
-        starknet_version: StarknetVersion::V0_13_4,
-        state_root: Felt::ZERO,
-        state_diff_commitment: Felt::ZERO,
-        transactions_commitment: Felt::ZERO,
-        receipts_commitment: Felt::ZERO,
-        events_commitment: Felt::ZERO,
-        transaction_count: 0,
-        events_count: 0,
-        state_diff_length: 0,
-    };
-
-    // the chain id is irrelevant here because the block is using starknet version 0.13.4
-    // only block pre 0.7.0 uses chain id in the hash computation
-    let hash = compute_hash(&header, &ChainId::SEPOLIA);
-
-    SealedBlockWithStatus {
-        block: SealedBlock { hash, header, body: Vec::new() },
-        status: FinalityStatus::AcceptedOnL2,
     }
 }
 
@@ -531,99 +416,4 @@ async fn downloaded_blocks_do_not_form_valid_chain() {
     // Verify no blocks were stored due to validation failure (except for block 99)
     let stored = get_stored_block_numbers(&provider, 99..=102);
     assert_eq!(stored.len(), 1);
-}
-
-#[tokio::test]
-async fn prune_compacts_state_history_at_boundary() {
-    let contract_address = ContractAddress::from(felt!("0x123"));
-    let class_hash: ClassHash = felt!("0xCAFE");
-    let nonce = felt!("0x7");
-    let storage_key = felt!("0x99");
-    let storage_value = felt!("0xBEEF");
-
-    // Only block 1 has state updates; later blocks are unchanged.
-    let mut updates_by_block = BTreeMap::new();
-    updates_by_block.insert(
-        1,
-        state_updates_with_contract_changes(
-            contract_address,
-            class_hash,
-            nonce,
-            storage_key,
-            storage_value,
-        ),
-    );
-
-    let provider = create_provider_with_block_range(0..=8, updates_by_block);
-    let mut stage =
-        katana_stage::IndexHistory::new(provider.clone(), TaskManager::current().task_spawner());
-
-    // keep_from = 5, prune range = [0, 5)
-    let output = stage.prune(&PruneInput::new(8, Some(3), None)).await.expect("prune must succeed");
-    assert!(output.pruned_count > 0, "expected prune to delete history entries");
-    assert_eq!(provider.provider_mut().earliest_available_state_block().unwrap(), Some(5));
-
-    // Historical reads in the retained window must still work due to anchor compaction.
-    let retained_state = provider.provider().historical(7.into()).unwrap().unwrap();
-    assert_eq!(retained_state.class_hash_of_contract(contract_address).unwrap(), Some(class_hash));
-    assert_eq!(retained_state.nonce(contract_address).unwrap(), Some(nonce));
-    assert_eq!(retained_state.storage(contract_address, storage_key).unwrap(), Some(storage_value));
-
-    // Pruned historical state provider range should be unavailable.
-    assert!(matches!(
-        provider.provider().historical(2.into()),
-        Err(ProviderError::HistoricalStatePruned { requested: 2, earliest_available: 5 })
-    ));
-
-    // Canonical per-block diffs remain exact even after history compaction.
-    assert_eq!(
-        provider.provider().state_update(1.into()).unwrap(),
-        Some(
-            state_updates_with_contract_changes(
-                contract_address,
-                class_hash,
-                nonce,
-                storage_key,
-                storage_value,
-            )
-            .state_updates
-        )
-    );
-    assert_eq!(provider.provider().state_update(5.into()).unwrap(), Some(StateUpdates::default()));
-}
-
-#[tokio::test]
-async fn historical_returns_pruned_error_below_retention_boundary() {
-    let provider = create_provider_with_block_range(0..=8, BTreeMap::new());
-
-    let provider_mut = provider.provider_mut();
-    provider_mut.set_earliest_available_state_block(5).unwrap();
-    provider_mut.commit().unwrap();
-
-    assert!(matches!(
-        provider.provider().historical(4.into()),
-        Err(ProviderError::HistoricalStatePruned { requested: 4, earliest_available: 5 })
-    ));
-    assert!(matches!(
-        provider.provider().historical(3.into()),
-        Err(ProviderError::HistoricalStatePruned { requested: 3, earliest_available: 5 })
-    ));
-    assert!(provider.provider().historical(5.into()).unwrap().is_some());
-    assert!(provider.provider().historical(8.into()).unwrap().is_some());
-}
-
-#[tokio::test]
-async fn index_history_prune_does_not_decrease_existing_retention_boundary() {
-    let provider = create_provider_with_block_range(0..=8, BTreeMap::new());
-    let mut stage =
-        katana_stage::IndexHistory::new(provider.clone(), TaskManager::current().task_spawner());
-
-    let provider_mut = provider.provider_mut();
-    provider_mut.set_earliest_available_state_block(10).unwrap();
-    provider_mut.commit().unwrap();
-
-    // keep_from=5
-
-    stage.prune(&PruneInput::new(8, Some(3), None)).await.expect("prune must succeed");
-    assert_eq!(provider.provider_mut().earliest_available_state_block().unwrap(), Some(10));
 }
